pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Personal AI Employee - Autonomous FTE System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=3.0.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.100.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.7.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.0",
    "mysqlclient>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Packages to include in the wheel
packages = ["database", "watchers", "skills"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.black]
line-length = 100
target-version = ['py313']
